Revert previous change that pinned setuptools to 62.0.0
As an attempt to get some of the recent PRs to build, it was necessary to pin setuptools to 62.0.0. In the most recent version (62.1.0) the locations of the build directories changed, causing compilation errors. Though this helped, the PR still failed...

I have fixed the setup.py file to deal with both the old 62.0.0 and 62.1.0 build directory locations, and will be starting a new release soon (today?), so it is no longer necessary to pin setuptools. We will see what happens then...